refactor: colocate code sample files in routes root
