+++ src/routes/abbreviations.svelte
@@ -4,7 +4,7 @@
 	import Example from './example.svelte';
 	import ValueDisplay from './value-display.svelte';
 	import { INPUT_CLASS } from './styles';
-	import code from './code-samples/abbreviations.md?raw';
+	import code from './abbreviations.md?raw';
 
 	let value = $state('');
 	let values = $state<CurrencyInputValues>({
