+++ src/routes/basic-usage.svelte
@@ -4,7 +4,7 @@
 	import Example from './example.svelte';
 	import ValueDisplay from './value-display.svelte';
 	import { INPUT_CLASS } from './styles';
-	import code from './code-samples/basic-usage.md?raw';
+	import code from './basic-usage.md?raw';
 
 	let value = $state('1234.56');
 	let values = $state<CurrencyInputValues>({
