+++ src/routes/decimal-precision.svelte
@@ -4,7 +4,7 @@
 	import Example from './example.svelte';
 	import ValueDisplay from './value-display.svelte';
 	import { INPUT_CLASS } from './styles';
-	import code from './code-samples/decimal-precision.md?raw';
+	import code from './decimal-precision.md?raw';
 
 	let value = $state('99.99');
 	let values = $state<CurrencyInputValues>({
