+++ src/routes/international-currencies.svelte
@@ -4,7 +4,7 @@
 	import Example from './example.svelte';
 	import ValueDisplay from './value-display.svelte';
 	import { INPUT_CLASS } from './styles';
-	import code from './code-samples/international-currencies.md?raw';
+	import code from './international-currencies.md?raw';
 
 	let value = $state('1234.56');
