+++ src/routes/tip-calculator.svelte
@@ -2,7 +2,7 @@
 	import { CurrencyInput, formatValue } from '$lib/index';
 	import Example from './example.svelte';
 	import { INPUT_CLASS } from './styles';
-	import code from './code-samples/tip-calculator.md?raw';
+	import code from './tip-calculator.md?raw';
 
 	const intlConfig = { locale: 'es-AR', currency: 'ARS' };
